Move Import redis to Core
