+++ tests/Wallabag/CoreBundle/Import/PocketHtmlImportTest.php
@@ -15,8 +15,8 @@ use Wallabag\CoreBundle\Entity\User;
 use Wallabag\CoreBundle\Helper\ContentProxy;
 use Wallabag\CoreBundle\Helper\TagsAssigner;
 use Wallabag\CoreBundle\Import\PocketHtmlImport;
+use Wallabag\CoreBundle\Redis\Producer;
 use Wallabag\CoreBundle\Repository\EntryRepository;
-use Wallabag\ImportBundle\Redis\Producer;
 
 class PocketHtmlImportTest extends TestCase
 {
